feat(ssr): server in typescript
