+++ Blog.Web/angular/package.json
@@ -3,9 +3,10 @@
   "version": "1.0.0",
   "scripts": {
     "lint": "tslint -p tsconfig.json",
-    "build": "npm run build:client && npm run build:server",
+    "build": "npm run build:client && npm run build:server && npm run webpack:server",
     "build:server": "ng run angular:server",
     "build:client": "ng build --prod",
+    "webpack:server": "webpack --config webpack.server.config.js --progress --colors",
     "clean": "rimraf wwwroot/dist clientapp/dist",
     "start": "ng serve & "
   },
@@ -39,14 +40,16 @@
     "zone.js": "0.8.26"
   },
   "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.0",
     "@angular/cli": "6.0.7",
     "@angular/service-worker": "6.0.3",
     "@types/chai": "^3.4.34",
     "@types/jasmine": "^2.8.4",
     "@types/node": "^9.4.6",
     "codelyzer": "^4.0.2",
+    "ts-loader": "^4.3.1",
     "tslint": "^5.9.1",
     "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "~0.6.0"
+    "webpack-cli": "^3.0.1"
   }
 }
