+++ Blog.Web/angular/webpack.server.config.js
@@ -1,0 +1,35 @@
+const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+  entry: { server: './server.ts' },
+  resolve: { extensions: ['.js', '.ts'] },
+  target: 'node',
+  mode: 'none',
+  // this makes sure we include node_modules and other 3rd party libraries
+  // externals: [/node_modules/],
+  output: {
+    path: path.join(__dirname, 'dist-server'),
+    filename: '[name].js',
+    libraryTarget: 'commonjs'
+  },
+  module: {
+    rules: [{ test: /\.ts$/, loader: 'ts-loader' }]
+  },
+  plugins: [
+    // fix for error "require is not defined"
+    new webpack.DefinePlugin({ 'global.GENTLY': false }),
+    // Temporary Fix for issue: https://github.com/angular/angular/issues/11580
+    // for 'WARNING Critical dependency: the request of a dependency is an expression'
+    new webpack.ContextReplacementPlugin(
+      /(.+)?angular(\\|\/)core(.+)?/,
+      path.join(__dirname, 'src'), // location of your src
+      {} // a map of your routes
+    ),
+    new webpack.ContextReplacementPlugin(
+      /(.+)?express(\\|\/)(.+)?/,
+      path.join(__dirname, 'src'),
+      {}
+    )
+  ]
+};
